Replace float with double

+++ GameDesignProject/Assets/Scripts/RecognizerScripts/Point.cs
@@ -6,10 +6,10 @@ namespace Recognizer
 {
     public class Point
     {
-        public float X { get; set; }
-        public float Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
     
-        public Point(float x, float y)
+        public Point(double x, double y)
         {
             X = x;
             Y = y;
